+++ packages/dendron-next-server/pages/vscode/note-preview.tsx
@@ -3,11 +3,12 @@ import {
   AssetGetThemeRequest,
   IntermediateDendronConfig,
   DMessageSource,
-  NoteViewMessageType,
   ThemeTarget,
   ThemeType,
   FOOTNOTE_REF_CLASS,
   FOOTNOTE_DEF_CLASS,
+  NoteViewMessageType,
+  NoteViewMessageEnum,
 } from "@dendronhq/common-all";
 import {
   createLogger,
@@ -212,7 +213,7 @@ function Note({ engine, ide, ws, port }: DendronProps & WorkspaceProps) {
         event.preventDefault();
         event.stopPropagation();
         postVSCodeMessage({
-          type: NoteViewMessageType.onClick,
+          type: NoteViewMessageEnum.onClick,
           data: {
             href: target.href,
             id: noteId,