+++ packages/dendron-plugin-views/src/views/DendronNotePageView.tsx
@@ -1,0 +1,7 @@
+import { renderOnDOM } from "../bootstrap";
+import DendronNotePage from "../components/DendronNotePage";
+
+renderOnDOM(DendronNotePage)
+
+// dummy export
+export {};