+++ packages/plugin-core/src/commands/ShowPreview.ts
@@ -1,19 +1,23 @@
 import {
   assertUnreachable,
   DendronWebViewKey,
-  DMessageType,
+  DMessageEnum,
   DNoteAnchor,
+  getStage,
   NoteProps,
+  NoteUtils,
   NoteViewMessage,
-  NoteViewMessageType,
+  NoteViewMessageEnum,
   OnDidChangeActiveTextEditorMsg,
+  VaultUtils,
 } from "@dendronhq/common-all";
+import { findUpTo, WebViewCommonUtils } from "@dendronhq/common-server";
 import _ from "lodash";
+import path from "path";
 import * as vscode from "vscode";
 import { DENDRON_COMMANDS } from "../constants";
 import { Logger } from "../logger";
-import { VSCodeUtils } from "../utils";
-import { WebViewUtils } from "../views/utils";
+import { VSCodeUtils, WSUtils } from "../utils";
 import { getEngine, getExtension } from "../workspace";
 import { BasicCommand } from "./base";
 import { GotoNoteCommand } from "./GotoNote";
@@ -62,16 +66,11 @@ const tryGetNoteFromDocument = (
   return;
 };
 
-/** Returns true when we are attempting to reference a place within the same note.
- *  Such as a different header section.*/
-const isLinkingWithinSameNote = (opts: {
-  data: { id?: string; href?: string };
-}) => {
-  const linkToNoteId = extractNoteIdFromHref(opts.data);
-
-  return linkToNoteId === opts.data.id;
-};
-
+/**
+ * Extract the last part of the path component
+ * @param data
+ * @returns
+ */
 export const extractNoteIdFromHref = (data: {
   id?: string;
   href?: string;
@@ -86,20 +85,86 @@ export const extractNoteIdFromHref = (data: {
   if (data.href.includes("vscode/note-preview")) {
     return data.id;
   }
+  /**
+   * Will return a link like 'vscode-webview://4e98b9cf-41d8-49eb-b458-fcfda32c6c01/foobar'
+   * The path component includes a `/` (eg. `/foobar`) so we remove it
+   */
+  const { path: hrefPath } = vscode.Uri.parse(data.href);
+  const out = path.basename(hrefPath, ".html");
+  return out;
+};
 
-  // In normal cases such as markdown value='[[#head2]]' with
-  // href='http://localhost:3005/vscode/0TDNEYgYvCs3ooZEuknNZ.html#head2' we will
-  // parse the href and extract the note id.
-  // The id within href can be without html suffix such as:
-  // * http://localhost:3005/vscode/0TDNEYgYvCs3ooZEuknNZ#head2
-  // * http://localhost:3005/vscode/0TDNEYgYvCs3ooZEuknNZ
-  //
-  // Regex is in reference to uuid.ts/genUUID() to match the note id.
-  // Regex: https://regex101.com/r/7pDj6G/1
-  const { path } = vscode.Uri.parse(data.href);
-  const noteId = path.match(/vscode\/([a-zA-Z0-9-_]*)/)?.[1];
+enum LinkType {
+  WIKI = "WIKI",
+  WEBSITE = "WEBSITE",
+  MARKDOWN = "MARKDOWN",
+  UNKNOWN = "UNKNOWN",
+}
 
-  return noteId;
+const classifyLink = ({ href }: NoteViewMessage["data"]): LinkType => {
+  if (href && href.startsWith("vscode-webview")) {
+    return LinkType.WIKI;
+  }
+  if (href && (href.startsWith("http://") || href.startsWith("https://"))) {
+    return LinkType.WEBSITE;
+  }
+  return LinkType.UNKNOWN;
+};
+
+const handleLink = async ({
+  linkType,
+  data,
+}: {
+  linkType: LinkType;
+  data: NoteViewMessage["data"];
+}) => {
+  switch (linkType) {
+    case LinkType.WIKI: {
+      // wiki links will have the following format
+      //
+      // with `prettyLinks` set to false
+      //    with anchor: vscode-webview://4e98b9cf-41d8-49eb-b458-fcfda32c6c01/foo.html'
+      //
+      // with `prettyLinks` set to true
+      //    with anchor: vscode-webview://4e98b9cf-41d8-49eb-b458-fcfda32c6c01/foo'
+      //    without anchor: vscode-webview://4e98b9cf-41d8-49eb-b458-fcfda32c6c01/foo#foobar'
+      //
+      // TODO: should be error
+      if (!data.href) return;
+      const noteId = extractNoteIdFromHref(data);
+      // TODO: should be error
+      if (!noteId) return;
+
+      const anchor = extractHeaderAnchorIfExists(data.href);
+      const note: NoteProps | undefined = getEngine().notes[noteId];
+      return new GotoNoteCommand().execute({
+        qs: note.fname,
+        vault: note.vault,
+        column: vscode.ViewColumn.One,
+        anchor,
+      });
+    }
+    case LinkType.WEBSITE: {
+      return VSCodeUtils.openLink(data.href!);
+    }
+    case LinkType.MARKDOWN: {
+      // assume local note - open relative to current vault
+      const note = getEngine().notes[data.id!];
+      if (!note) return;
+
+      // noteId in this case is name of file
+      const fullPath = path.join(
+        getEngine().wsRoot,
+        VaultUtils.getRelPath(note.vault),
+        note.id
+      );
+      return VSCodeUtils.openLink(fullPath);
+    }
+    default: {
+      // default, do nothing. let vscode link handler handle it
+      return;
+    }
+  }
 };
 
 export class ShowPreviewCommand extends BasicCommand<
@@ -121,7 +186,7 @@ export class ShowPreviewCommand extends BasicCommand<
       Logger.debug({ ...ctx, state: "panel found" });
       panel.title = `${title} ${note.fname}`;
       panel.webview.postMessage({
-        type: DMessageType.ON_DID_CHANGE_ACTIVE_TEXT_EDITOR,
+        type: DMessageEnum.ON_DID_CHANGE_ACTIVE_TEXT_EDITOR,
         data: {
           note,
           syncChangedNote: true,
@@ -139,28 +204,38 @@ export class ShowPreviewCommand extends BasicCommand<
   }
 
   async execute(_opts?: CommandOpts) {
-    // Get workspace information
+    const ctx = "ShowPreview";
     const ext = getExtension();
-
-    // If panel already exists
     const existingPanel = ext.getWebView(DendronWebViewKey.NOTE_PREVIEW);
-
     const viewColumn = vscode.ViewColumn.Beside; // Editor column to show the new webview panel in.
     const preserveFocus = true;
 
+    // show existing panel if exist
     if (!_.isUndefined(existingPanel)) {
+      Logger.info({ ctx, msg: "reveal existing" });
       try {
         // If error, panel disposed and needs to be recreated
         existingPanel.reveal(viewColumn, preserveFocus);
         return;
-      } catch (error) {
-        console.error(error);
+      } catch (error: any) {
+        Logger.error({ ctx, error });
       }
     }
+    Logger.info({ ctx, msg: "creating new" });
 
+    const assetUri = WSUtils.getAssetUri(getExtension().context);
+    const pkgRoot = path.dirname(
+      findUpTo({ base: __dirname, fname: "package.json", maxLvl: 5 })!
+    );
+    const pluginViewsRoot: vscode.Uri =
+      getStage() === "dev"
+        ? vscode.Uri.file(
+            path.join(pkgRoot, "..", "dendron-plugin-views", "build")
+          )
+        : assetUri;
     const panel = vscode.window.createWebviewPanel(
-      "dendronIframe", // Identifies the type of the webview. Used internally
-      title, // Title of the panel displayed to the user
+      "dendronPreview",
+      "Dendron Preview",
       {
         viewColumn,
         preserveFocus,
@@ -169,52 +244,53 @@ export class ShowPreviewCommand extends BasicCommand<
         enableScripts: true,
         retainContextWhenHidden: true,
         enableFindWidget: true,
-        enableCommandUris: true,
+        localResourceRoots: [assetUri, pluginViewsRoot],
       }
     );
 
-    const resp = await WebViewUtils.genHTMLForWebView({
-      title,
-      view: DendronWebViewKey.NOTE_PREVIEW,
-    });
-
-    panel.webview.html = resp;
-
+    const name = "notePreview";
+    const jsSrc = vscode.Uri.joinPath(
+      pluginViewsRoot,
+      "static",
+      "js",
+      `${name}.bundle.js`
+    );
+    const cssSrc = vscode.Uri.joinPath(
+      pluginViewsRoot,
+      "static",
+      "css",
+      `${name}.styles.css`
+    );
+    const port = getExtension().port!;
     panel.webview.onDidReceiveMessage(async (msg: NoteViewMessage) => {
       const ctx = "ShowPreview:onDidReceiveMessage";
+      Logger.debug({ ctx, msgType: msg.type });
       switch (msg.type) {
-        case NoteViewMessageType.onClick: {
-          const { data } = msg;
-          if (data.href) {
-            // TODO find a better way to differentiate local files from web links (`data-` attribute)
-            if (data.href.includes("localhost")) {
-              const noteId = extractNoteIdFromHref(data);
-              const anchor = extractHeaderAnchorIfExists(data.href);
-
-              const note: NoteProps | undefined = this.getNote(noteId);
-              if (isLinkingWithinSameNote({ data }) && note && anchor) {
-                await new GotoNoteCommand().execute({
-                  qs: note.fname,
-                  vault: note.vault,
-                  column: vscode.ViewColumn.One,
-                  anchor,
-                });
-              } else if (noteId && note) {
-                await new GotoNoteCommand().execute({
-                  qs: note.fname,
-                  vault: note.vault,
-                  column: vscode.ViewColumn.One,
-                  anchor,
-                });
-              }
-            } else {
-              VSCodeUtils.openLink(data.href);
-            }
-          }
-
+        case DMessageEnum.ON_DID_CHANGE_ACTIVE_TEXT_EDITOR:
+        case DMessageEnum.INIT: {
+          // do nothing
           break;
         }
-        case NoteViewMessageType.onGetActiveEditor: {
+        case DMessageEnum.MESSAGE_DISPATCHER_READY: {
+          // if ready, get current note
+          const note = VSCodeUtils.getActiveNote();
+          if (note) {
+            Logger.debug({
+              ctx,
+              msg: "got active note",
+              note: NoteUtils.toLogObj(note),
+            });
+            ShowPreviewCommand.refresh(note);
+          }
+          break;
+        }
+        case NoteViewMessageEnum.onClick: {
+          const { data } = msg;
+          const linkType = classifyLink(data);
+          await handleLink({ linkType, data });
+          break;
+        }
+        case NoteViewMessageEnum.onGetActiveEditor: {
           // only entered on "init" in `plugin-core/src/views/utils.ts:87`
           Logger.debug({ ctx, "msg.type": "onGetActiveEditor" });
           const activeTextEditor = VSCodeUtils.getActiveTextEditor();
@@ -224,14 +300,22 @@ export class ShowPreviewCommand extends BasicCommand<
           if (!_.isUndefined(maybeNote)) ShowPreviewCommand.refresh(maybeNote);
           break;
         }
-        case NoteViewMessageType.messageDispatcherReady:
-          break;
         default:
           assertUnreachable(msg.type);
       }
     });
 
-    // Update workspace-wide graph panel
+    const html = getWebviewContent({
+      jsSrc,
+      cssSrc,
+      port,
+      wsRoot: ext.getEngine().wsRoot,
+      panel,
+    });
+
+    panel.webview.html = html;
+
+    // Update workspace-wide panel
     ext.setWebView(DendronWebViewKey.NOTE_PREVIEW, panel);
 
     // remove webview from workspace when user closes it
@@ -242,11 +326,39 @@ export class ShowPreviewCommand extends BasicCommand<
       ext.setWebView(DendronWebViewKey.NOTE_PREVIEW, undefined);
     });
   }
+}
 
-  private getNote(noteId: string | undefined) {
-    if (noteId === undefined) {
-      return undefined;
-    }
-    return getEngine().notes[noteId];
-  }
+function getWebviewContent({
+  jsSrc,
+  cssSrc,
+  port,
+  wsRoot,
+  panel,
+}: {
+  jsSrc: vscode.Uri;
+  cssSrc: vscode.Uri;
+  port: number;
+  wsRoot: string;
+  panel: vscode.WebviewPanel;
+}) {
+  const root = WSUtils.getAssetUri(getExtension().context);
+  const themes = ["light", "dark"];
+  const themeMap: any = {};
+  themes.map((th) => {
+    themeMap[th] = panel.webview
+      .asWebviewUri(
+        vscode.Uri.joinPath(root, "static", "css", "themes", `${th}.css`)
+      )
+      .toString();
+  });
+  const out = WebViewCommonUtils.genVSCodeHTMLIndex({
+    jsSrc: panel.webview.asWebviewUri(jsSrc).toString(),
+    cssSrc: panel.webview.asWebviewUri(cssSrc).toString(),
+    port,
+    wsRoot,
+    browser: false,
+    acquireVsCodeApi: `const vscode = acquireVsCodeApi(); window.vscode = vscode;`,
+    themeMap,
+  });
+  return out;
 }